tech-feature-minor-refactor model holder service added. model service code is now using model holder service and internal data structure details are moved to holder service.
